fix: markdown and separate connector main component
